Add files via upload
The text can be obtained from any online source where scrapping is easily possible. This code has been specifically tested for Wikipedia pages. After obtaining the text, all unwanted punctuations and stop words are removed once the function word_tokenized has been applied to the text. After the cleaning of the worst, a word frequency dictionary is made to find the weighted frequency of each word. A list of all the sentences in the text obtained is made using the function sent_tokenize. Sentence scores are calculated by summing the weighted frequencies of the words found in the respective sentence, which is added to the dictionary sentence_scores. The top 7 sentences with the highest scores have then been joined to give us the summary.